Alterações Aula 11/03 Website com listagem

+++ Clima/Models/Tb_registos.cs
@@ -7,8 +7,8 @@ namespace Clima.Models
     {
         [Key]
         public int ID_registo { get; set; }
-        public int temperatura { get; set; }
-        public int humidade { get; set; }
+        public double temperatura { get; set; }
+        public double humidade { get; set; }
         public int risco_temperatura { get; set; }
         public int risco_humidade { get; set; }
         public DateTime data_registo { get; set; } = DateTime.Now;
